export correctly
