+++ packages/thirdweb/src/exports/indexer.ts
@@ -1,1 +1,27 @@
-export * from "../indexer/index.js";
+export {
+  getBlock,
+  type GetBlockParams,
+} from "../indexer/endpoints/getBlock.js";
+export {
+  getEvents,
+  type GetEventsParams,
+  type GetEventsInterval,
+  type GetEventsGroupBy,
+} from "../indexer/endpoints/getEvents.js";
+export {
+  getLatestBlockNumber,
+  type GetLatestBlockNumberParams,
+} from "../indexer/endpoints/getLatestBlockNumber.js";
+export {
+  getNFTsByCollection,
+  type GetNFTsByCollectionParams,
+  type GetNFTsGroupBy,
+} from "../indexer/endpoints/getNFTsByCollection.js";
+export {
+  getNFTsByOwner,
+  type GetNFTsByOwnerParams,
+} from "../indexer/endpoints/getNFTsByOwner.js";
+export {
+  getTransactions,
+  type GetTransactionsParams,
+} from "../indexer/endpoints/getTransactions.js";
