--- packages/thirdweb/src/indexer/index.ts
@@ -1,24 +1,0 @@
-export { getBlock, type GetBlockParams } from "./endpoints/getBlock.js";
-export {
-  getEvents,
-  type GetEventsParams,
-  type GetEventsInterval,
-  type GetEventsGroupBy,
-} from "./endpoints/getEvents.js";
-export {
-  getLatestBlockNumber,
-  type GetLatestBlockNumberParams,
-} from "./endpoints/getLatestBlockNumber.js";
-export {
-  getNFTsByCollection,
-  type GetNFTsByCollectionParams,
-  type GetNFTsGroupBy,
-} from "./endpoints/getNFTsByCollection.js";
-export {
-  getNFTsByOwner,
-  type GetNFTsByOwnerParams,
-} from "./endpoints/getNFTsByOwner.js";
-export {
-  getTransactions,
-  type GetTransactionsParams,
-} from "./endpoints/getTransactions.js";
